LPS-76025 split registry into interface and impl, move interface and util to api module for export
